+++ src/main/java/com/belhard/bookstore/dao/user/UserDaoImpl.java
@@ -53,7 +53,7 @@ public class UserDaoImpl implements UserDao {
     }
 
     @Override
-    public User read(int id) {
+    public User read(Long id) {
         try (Connection connection = dataSource.getConnection()) {
             logger.debug("Reading user", id);
             PreparedStatement statement = connection.prepareStatement(SELECT_QUERY);
